Moved date functions into lib/dates.js.
